+++ packages/element-controller/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @open-cells/element-controller
 
+## 1.0.2-rc.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @open-cells/core-plugin@1.2.0-rc.0
+
 ## 1.0.1
 
 ### Patch Changes