+++ packages/example/blank-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # blank-app
 
+## 0.1.2-rc.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @open-cells/core@1.1.0-rc.0
+  - @open-cells/element-controller@1.0.2-rc.0
+  - @open-cells/page-controller@1.0.2-rc.0
+
 ## 0.1.1
 
 ### Patch Changes